+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - TODO: Make GUI behave like normal inventory (let people take out items manually)
 - TODO: Fix logging, see experimental branch
 
+## 13.8.0
+- Fixed teleporting not properly checking for nearby safe locations and instead teleporting the player to the chest location
+- Added two undocumented settings to config (tp-distance, tp-no-safe-space-y-offset)
+
 ## 13.7.0
 - fixed-chest-offset now has a "relative mode" (see config.yml)
 - Oraxen furniture yaw and face can now be set in config.yml (see "oraxen-furniture-yaw" and "oraxen-furniture-face" in config.yml)